Reduce RelativeArchivePath and ArchiveFileTreeElement
